OpenAPI_automationTesting/requirements.txt
allure-pytest==2.6.1
allure-python-commons==2.6.1
atomicwrites==1.2.1
attrs==18.2.0
beautifulsoup4==4.6.3
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
enum34==1.1.6
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4
jieba==0.39
kiwisolver==1.0.1
lxml==4.3.2
matplotlib==2.2.2
more-itertools==4.3.0
Naked==0.1.31
namedlist==1.7
numpy==1.15.0
openpyxl==2.5.7
Pillow==5.2.0
pluggy==0.8.0
py==1.7.0
pyparsing==2.2.0
pytest==4.0.2
pytest-logging==2015.11.4
pytest-rerunfailures==6.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.21.0
selenium==3.14.0
shellescape==3.4.1
simplejson==3.16.0
six==1.12.0
suds-py3==1.3.3.0
urllib3==1.24.1
wordcloud==1.4.1
xlrd==1.1.0
